requirements.txt
fastmcp>=0.1.0
playwright>=1.40.0
PyGithub>=2.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
